Moving header theme files into a subfolder
