+++ src/theme/index.ts
@@ -20,15 +20,15 @@ import { CustomImage, CustomImageWrapper } from "./components/image";
 import CustomSelect from "./components/select";
 import DatePicker from "./components/datePicker";
 import Fieldset from "./components/fieldset";
-import Header from "./components/header";
-import HeaderLogin from "./components/headerLogin";
-import HeaderLoginButton from "./components/headerLoginButton";
-import HeaderLowerNav from "./components/headerLowerNav";
-import HeaderMobileNav from "./components/headerMobileNav";
-import HeaderMobileNavButton from "./components/headerMobileNavButton";
-import HeaderSearchForm from "./components/headerSearchForm";
-import HeaderSearchButton from "./components/headerSearchButton";
-import HeaderUpperNav from "./components/headerUpperNav";
+import Header from "./components/header/header";
+import HeaderLogin from "./components/header/headerLogin";
+import HeaderLoginButton from "./components/header/headerLoginButton";
+import HeaderLowerNav from "./components/header/headerLowerNav";
+import HeaderMobileNav from "./components/header/headerMobileNav";
+import HeaderMobileNavButton from "./components/header/headerMobileNavButton";
+import HeaderSearchForm from "./components/header/headerSearchForm";
+import HeaderSearchButton from "./components/header/headerSearchButton";
+import HeaderUpperNav from "./components/header/headerUpperNav";
 import Heading from "./components/heading";
 import HelperErrorText from "./components/helperErrorText";
 import Hero from "./components/hero";
